simple version

+++ simple/tetris.js
@@ -9,25 +9,29 @@ class Tetris {
             nextCanvas:'next', 
             cols:10, 
             rows:20,
-            use7Bag: true
+            use7Bag: true,
+            showNextPiece: true,
+            increaseLevel: true,
+            onGameOver: null,
+            onBeforeConnect:null,
+            onLineClear: null,
+            onStartTurn:null,
+            onReset: null
         }, options);
-        
-        this.scale = this.options.scale;
-        this.tickSpeed = this.options.tickSpeed;
 
         this.canvas = document.getElementById(this.options.gameCanvas);
         if (this.canvas) {
-            this.canvas.width = this.options.cols * this.scale;
-            this.canvas.height = this.options.rows * this.scale;
+            this.canvas.width = this.options.cols * this.options.scale;
+            this.canvas.height = this.options.rows * this.options.scale;
             this.ctx = this.canvas.getContext('2d');
-            this.ctx.scale(this.scale, this.scale);
+            this.ctx.scale(this.options.scale, this.options.scale);
         }
         this.nextCanvas = document.getElementById(this.options.nextCanvas);
         if (this.nextCanvas) {
-            this.nextCanvas.width = 4 * this.scale;
-            this.nextCanvas.height = 4 * this.scale;
+            this.nextCanvas.width = 4 * this.options.scale;
+            this.nextCanvas.height = 4 * this.options.scale;
             this.nextCtx = this.nextCanvas.getContext('2d');
-            this.nextCtx.scale(this.scale, this.scale);
+            this.nextCtx.scale(this.options.scale, this.options.scale);
         }
 
         this.grid = this.createMatrix(10, this.options.rows + 2);
@@ -52,19 +56,58 @@ class Tetris {
         this.dropCounter = 0;
         this.dropInterval = 500;
 
+        this.running = false;
+
         this.bag = ['T', 'O', 'L', 'J', 'I', 'S', 'Z'];
         this.bag.sort(() => Math.random() - 0.5);
 
         this.draw = this.draw.bind(this);
         this.drawNextPiece = this.drawNextPiece.bind(this);
         this.update = this.update.bind(this);
-        if (this.canvas) this.draw();
         if (this.nextCanvas) this.drawNextPiece();
+        if (this.canvas) {
+            this.draw();
+            document.addEventListener('keydown', event => {
+                if (!this.running) return;
+                if (event.keyCode === 32) this.drop(Infinity);
+                else if (event.keyCode === 37) this.move(-1);
+                else if (event.keyCode === 39) this.move(1);
+                else if (event.keyCode === 38)  this.rotate(1);
+                else if (event.keyCode === 40) this.drop();
+            });
+        }
+        this.update();
     }
-
+    
     start() {
         this.pieceReset();
-        this.update();
+        this.running = true;
+        this.options.onReset?.();
+    }
+
+    gameOver() {
+        this.running = false;
+        this.options.onGameOver?.();
+    }
+
+    reset() {
+        this.grid = this.createMatrix(this.options.cols, this.options.rows + 2);
+        this.piece = null;
+        this.nextPiece = null;
+        this.points = 0;
+        this.lines = 0;
+        this.level = 1;
+        this.dropCounter = 0;
+        this.dropInterval = 500;
+        this.bag = ['T', 'O', 'L', 'J', 'I', 'S', 'Z'];
+        this.bag.sort(() => Math.random() - 0.5);
+        if (this.canvas) {
+            this.canvas.width = this.options.cols * this.options.scale;
+            this.canvas.height = this.options.rows * this.options.scale;
+            this.ctx = this.canvas.getContext('2d');
+            this.ctx.scale(this.options.scale, this.options.scale);
+        }
+        this.start();
     }
 
     createMatrix(w, h) {
@@ -88,7 +131,8 @@ class Tetris {
         return false;
     }
 
-    merge(grid, piece) {
+    merge(piece, grid) {
+        this.options.onBeforeConnect?.(piece, grid);
         piece.matrix.forEach((row, y) => {
             row.forEach((value, x) => {
                 if (value !== 0) {
@@ -98,15 +142,21 @@ class Tetris {
         });
     }
 
-    drop(piece, grid) {
+    drop(amount, piece, grid) {
+        if (!this.running) return;
         if (!piece) piece = this.piece;
         if (!grid) grid = this.grid;
-        piece.pos.y++;
-        if (this.collide(grid, piece.matrix, piece.pos)) {
-            piece.pos.y--;
-            this.merge(grid, piece);
-            this.pieceReset();
-            this.clearLines();
+        if (!amount) amount = 1;
+        for (let i = 0; i < amount; i++)
+        {
+            piece.pos.y++;
+            if (this.collide(grid, piece.matrix, piece.pos)) {
+                piece.pos.y--;
+                this.merge(piece, grid);
+                this.pieceReset();
+                this.clearLines();
+                break;
+            }
         }
         this.dropCounter = 0;
     }
@@ -136,13 +186,16 @@ class Tetris {
     }
 
     pieceReset() {
-        
         this.piece = this.createPiece(this.nextPiece?.type || this.randomizePiece());
         this.nextPiece = this.createPiece(this.randomizePiece());
 
         if (this.collide(this.grid, this.piece.matrix, this.piece.pos)) {
-            this.grid.forEach(row => row.fill(0));
+            this.gameOver();
+            return false;
         }
+
+        this.options.onStartTurn?.(this.piece?.type, this.nextPiece?.type);
+        return true;
     }
 
     rotate(dir, piece, grid) {
@@ -176,10 +229,23 @@ class Tetris {
         }
     }
 
+    projectGhost(piece, grid) {
+        if (!piece) piece = this.piece;
+        if (!grid) grid = this.grid;
+        if (!piece || !grid) return;
+        const ghost = JSON.parse(JSON.stringify(piece));
+        while (!this.collide(grid, ghost.matrix, ghost.pos)) {
+            ghost.pos.y++;
+        }
+        ghost.pos.y--;
+        return ghost;
+    }
+
     drawNextPiece() {
         requestAnimationFrame(this.drawNextPiece);
         this.nextCtx.fillStyle = '#F0EAD6';
         this.nextCtx.fillRect(0, 0, this.nextCanvas.width, this.nextCanvas.height);
+        if (!this.options.showNextPiece) return;
         this.drawMatrix(this.nextCtx, this.nextPiece?.matrix, { 
             x: (4 - this.nextPiece?.matrix[0].length) / 2 , 
             y: (4 - this.nextPiece?.matrix.length) / 2
@@ -191,15 +257,16 @@ class Tetris {
         this.ctx.fillStyle = '#F0EAD6';
         this.ctx.fillRect(0, 0, this.canvas.width, this.canvas.height);
         this.drawMatrix(this.ctx, this.grid, { x: 0, y: -2 });
+        this.drawMatrix(this.ctx, this.projectGhost()?.matrix, { x: this.projectGhost()?.pos?.x, y: this.projectGhost()?.pos?.y - 2 }, '#00000022');
         this.drawMatrix(this.ctx, this.piece?.matrix, { x: this.piece?.pos?.x, y: this.piece?.pos?.y - 2 });
     }
 
-    drawMatrix(ctx, matrix, offset) {
+    drawMatrix(ctx, matrix, offset, colorOverride) {
         if (!matrix) return;
         matrix.forEach((row, y) => {
             row.forEach((value, x) => {
                 if (value !== 0) {
-                    ctx.fillStyle = this.colors[value];
+                    ctx.fillStyle = colorOverride ? colorOverride : this.colors[value];
                     ctx.fillRect(x + offset.x,
                         y + offset.y,
                         1, 1);
@@ -211,7 +278,8 @@ class Tetris {
     update() {
         //requestAnimationFrame(this.update);
         setTimeout(() => this.update(), 0);
-        for (let i = 0; i < this.tickSpeed; i++) {
+        if (!this.running) return;
+        for (let i = 0; i < this.options.tickSpeed; i++) {
             this.dropCounter += 10;
             if (this.dropCounter > this.dropInterval) {
                 this.drop();
@@ -225,18 +293,21 @@ class Tetris {
     }
 
     clearLines() {
+        let amount = 0;
         for (let y = this.grid.length - 1; y >= 0; y--) {
             let isLineFull = this.grid[y].every(value => value !== 0);
             if (isLineFull) {
                 this.grid.splice(y, 1);
                 this.grid.unshift(new Array(this.grid[0].length).fill(0));
                 this.points += 100;
-                this.lines++;
-                if (this.lines >= this.level * 10) {
+                amount++;
+                if (this.options.increaseLevel && this.lines >= this.level * 10) {
                     this.increaseLevel();
                 }
             }
         }
+        this.lines += amount;
+        this.options.onLineClear?.(amount);
     }
 
     createPiece(type) {
